+++ Makefile
@@ -3,6 +3,6 @@ test:
 	go test -v -race -cover ./plugins/rpc -tags=debug
 	go test -v -race -cover ./plugins/rpc/tests -tags=debug
 	go test -v -race -cover ./plugins/config/tests -tags=debug
-	go test -v -race -cover ./plugins/app/tests -tags=debug
+	go test -v -race -cover ./plugins/server/tests -tags=debug
 	go test -v -race -cover ./plugins/logger/tests -tags=debug
 	go test -v -race -cover ./plugins/metrics/tests -tags=debug
