+++ plugins/server/tests/configs/.rr-no-app-section.yaml
@@ -1,9 +1,9 @@
-app:
-  command: "php socket.php"
+server:
+  command: "php ../../../tests/client.php echo pipes"
   user: ""
   group: ""
   env:
     "RR_CONFIG": "/some/place/on/the/C134"
     "RR_CONFIG2": "C138"
-  relay: "unix://unix.sock"
+  relay: "pipes"
   relayTimeout: "20s"
