+++ plugins/server/tests/configs/.rr-sockets.yaml
@@ -1,9 +1,9 @@
-app:
-  command: "php some_absent_file.php"
+server:
+  command: "php socket.php"
   user: ""
   group: ""
   env:
     "RR_CONFIG": "/some/place/on/the/C134"
     "RR_CONFIG2": "C138"
-  relay: "pipes"
+  relay: "unix://unix.sock"
   relayTimeout: "20s"
