+++ plugins/server/tests/configs/.rr-tcp.yaml
@@ -1,9 +1,9 @@
-app:
-  command: "php ../../../tests/client.php echo pipes"
+server:
+  command: "php tcp.php"
   user: ""
   group: ""
   env:
     "RR_CONFIG": "/some/place/on/the/C134"
     "RR_CONFIG2": "C138"
-  relay: "pupes"
+  relay: "tcp://localhost:9999"
   relayTimeout: "20s"
