+++ plugins/server/tests/configs/.rr-wrong-relay.yaml
@@ -1,9 +1,9 @@
-app:
+server:
   command: "php ../../../tests/client.php echo pipes"
   user: ""
   group: ""
   env:
     "RR_CONFIG": "/some/place/on/the/C134"
     "RR_CONFIG2": "C138"
-  relay: "pipes"
+  relay: "pupes"
   relayTimeout: "20s"
